fix: scrollParent detection
