+++ src/dom-utils/getScrollParent.js
@@ -4,10 +4,7 @@ import getComputedStyle from './getComputedStyle';
 import getNodeName from './getNodeName';
 import { isHTMLElement } from './instanceOf';
 
-export default function getOverflowParent(
-  node: Node,
-  type: 'scroll' | 'clipping' = 'scroll'
-): HTMLElement {
+export default function getScrollParent(node: Node): HTMLElement {
   if (['html', 'body', '#document'].includes(getNodeName(node))) {
     return node.ownerDocument.body;
   }
@@ -15,13 +12,11 @@ export default function getOverflowParent(
   if (isHTMLElement(node)) {
     // Firefox wants us to check `-x` and `-y` variations as well
     const { overflow, overflowX, overflowY } = getComputedStyle(node);
-    const re = type === 'scroll' ? /visible|hidden/ : /visible/;
-    const str = overflow + overflowY + overflowX;
 
-    if (str !== '' && !re.test(str)) {
+    if (/auto|scroll|overlay|hidden/.test(overflow + overflowY + overflowX)) {
       return node;
     }
   }
 
-  return getOverflowParent(getParentNode(node), type);
+  return getScrollParent(getParentNode(node));
 }
