--- src/dom-utils/listClippingParents.js
@@ -1,6 +1,0 @@
-// @flow
-import listOverflowParents from './listOverflowParents';
-
-export default function listClippingParents(element: Node): Array<Element> {
-  return listOverflowParents(element, 'clipping');
-}
